pyproject.toml
[tool.poetry]
name = "llmchat-for-child-emotion"
version = "0.1.0"
description = ""
authors = ["Woosuk Seo & Young-Ho Kim"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.99.1"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
openai = "^0.27.8"
nanoid = "^2.0.0"
python-dotenv = "^1.0.0"
pendulum = "^2.1.2"
jsonlines = "^3.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
